Upload Reprogramable Array program
The program uses an array to determine which actions to take. You reprogram the array (starting from Pos 1, so skip the 0 spot, due to how loop index works, not my choice), with values that determine which action to take. 1 goes forward, 2 rotates right, 3 rotates left, 4 goes backwards, 5 stops the vehicle. The values are small intentionally, so that you may adjust as needed. The array is accessed, and the value pulled is from the loop index. The result is placed in the switch, and the operation is completed. The program waits three seconds before activating again. You can stop the program / loop in it's entirety with a value of six.

This has NOT been tested