Add Prototype Infrared Maze Program
Since I know we're having to do an infrared test, I'm just prototyping out what I think it'll look like before we begin the actual challenge. This definitely has a few kinks in the program so I wouldn't really call this ready to be used. It's more for my self practice, and so I can check later when we begin.