Add the Completed Line Follower
This Line Following program requires a color sensor to be pointed at the ground. It reads for a black piece of tape, and follows it to the end. It's rather simple, it just turns right when it sees black, and otherwise (when it would go off the line), it will turn left. It works, video will be placed here at some point.