Race Track Program - Controller / Driver
This file is actually two programs: The Controller, and the Driver. The driver needs an ultrasonic sensor and a color sensor, the controller needs a touch sensor and ultrasonic sensor. 

The Controller will transmit an ultrasonic signal until the button is pressed, where it pauses for a moment. The Driver will wait for the signal from the controller, where it will stop the program once it detects it. This can be used to remotely stop the robot.